+++ examples/remove-emphasis/mdbook-remove-emphasis/Cargo.toml
@@ -4,8 +4,7 @@ version = "0.1.0"
 edition.workspace = true
 
 [dependencies]
-anyhow.workspace = true
-mdbook = { path = "../../.." }
+mdbook-preprocessor.workspace = true
 pulldown-cmark = { version = "0.12.2", default-features = false }
 pulldown-cmark-to-cmark = "18.0.0"
 serde_json = "1.0.132"
