+++ tests/testsuite/preprocessor.rs
@@ -3,7 +3,8 @@
 use crate::prelude::*;
 use anyhow::Result;
 use mdbook::book::Book;
-use mdbook::preprocess::{CmdPreprocessor, Preprocessor, PreprocessorContext};
+use mdbook::preprocess::CmdPreprocessor;
+use mdbook_preprocessor::{Preprocessor, PreprocessorContext};
 use std::sync::{Arc, Mutex};
 
 struct Spy(Arc<Mutex<Inner>>);
